Moved FeelingLucky/Games/Tools Activities into Fragments
