+++ IceIceBB/app/src/main/AndroidManifest.xml
@@ -1,16 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.lmont.iceicebb">
+          package="com.example.lmont.iceicebb">
 
-    <uses-permission
-        android:name="android.permission.INTERNET"/>
-    <uses-permission
-        android:name="android.permission.READ_SYNC_SETTINGS"/>
-    <uses-permission
-        android:name="android.permission.WRITE_SYNC_SETTINGS"/>
-    <uses-permission
-        android:name="android.permission.AUTHENTICATE_ACCOUNTS"/>
-
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.READ_SYNC_SETTINGS"/>
+    <uses-permission android:name="android.permission.WRITE_SYNC_SETTINGS"/>
+    <uses-permission android:name="android.permission.AUTHENTICATE_ACCOUNTS"/>
     <uses-permission android:name="android.permission.WRITE_SYNC_SETTINGS"/>
 
     <application
@@ -21,65 +16,69 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.example.lmont.iceicebb.SplashScreenActivity"
+            android:name=".SplashScreenActivity"
             android:label="@string/app_name"
-            android:theme="@style/Splashscreen"
+            android:noHistory="true"
             android:screenOrientation="portrait"
-            android:noHistory="true" >
-
-
+            android:theme="@style/Splashscreen">
             <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+                <action android:name="android.intent.action.MAIN"/>
 
-                <category android:name="android.intent.category.LAUNCHER" />
+                <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
             <intent-filter>
-                <action android:name="android.intent.action.SEARCH" />
+                <action android:name="android.intent.action.SEARCH"/>
             </intent-filter>
-            <meta-data android:name="android.app.searchable"
-                android:resource="@xml/searchable" />
+
+            <meta-data
+                android:name="android.app.searchable"
+                android:resource="@xml/searchable"/>
         </activity>
-        <activity android:name=".GameActivity"
-
-            android:label="@string/app_name">
+        <activity
+            android:name=".TabMainActivity"
+            android:label="@string/title_activity_tab_main">
             <intent-filter>
+
+
                 <category android:name="android.intent.category.DEFAULT"/>
+
+
             </intent-filter>
-
-
         </activity>
-        <activity android:name=".GameDetailActivity" />
-        <activity android:name=".ToolsActivity" />
-        <activity android:name=".CardsActivity" />
-        <activity android:name=".DiceActivity" />
+        <activity
+            android:name=".GameActivity"
+            android:label="@string/app_name">
+        </activity>
+        <activity android:name=".GameDetailActivity"/>
+        <activity android:name=".ToolsActivity"/>
+        <activity android:name=".CardsActivity"/>
+        <activity android:name=".DiceActivity"/>
         <activity android:name=".FeelingLuckyActivity"/>
 
-
-        <service
-            android:name="com.example.lmont.iceicebb.PlaceholderAuthenticatorService">
+        <service android:name=".PlaceholderAuthenticatorService">
             <intent-filter>
                 <action android:name="android.accounts.AccountAuthenticator"/>
             </intent-filter>
+
             <meta-data
                 android:name="android.accounts.AccountAuthenticator"
-                android:resource="@xml/authenticator" />
+                android:resource="@xml/authenticator"/>
         </service>
-
-        <service
-            android:name="com.example.lmont.iceicebb.MySyncService">
+        <service android:name=".MySyncService">
             <intent-filter>
                 <action android:name="android.content.SyncAdapter"/>
             </intent-filter>
+
             <meta-data
                 android:name="android.content.SyncAdapter"
-                android:resource="@xml/syncadapter" />
+                android:resource="@xml/syncadapter"/>
         </service>
 
         <provider
             android:name=".IcebreakerContentProvider"
             android:authorities="com.example.lmont.iceicebb.IcebreakerContentProvider"
-            android:label="my provider"
             android:exported="false"
+            android:label="my provider"
             android:syncable="true"/>
