+++ IceIceBB/app/src/main/java/com/example/lmont/iceicebb/Fragments/ToolsFragment.java
@@ -12,7 +12,7 @@ import com.example.lmont.iceicebb.R;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class MaterialFragment extends Fragment {
+public class ToolsFragment extends Fragment {
 
 
 
@@ -21,10 +21,10 @@ public class MaterialFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_material, container, false);
+        View view = inflater.inflate(R.layout.fragment_tools, container, false);
 
 
-        return null;
+        return view;
     }
 
 }
