+++ IceIceBB/app/src/main/res/layout/activity_game.xml
@@ -7,16 +7,10 @@
                 android:layout_centerHorizontal="true"
                 android:background="#587e7c"
                 android:paddingBottom="@dimen/activity_vertical_margin"
-                android:paddingLeft="@dimen/activity_horizontal_margin"
-                android:paddingRight="@dimen/activity_horizontal_margin"
-                android:paddingTop="2dp"
+                android:paddingTop="1dp"
                 tools:context="com.example.lmont.iceicebb.GameActivity">
 
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/gameList"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_centerHorizontal="true"/>
+
 
     <android.support.design.widget.TabLayout
         android:id="@+id/tab_layout"
@@ -32,42 +26,6 @@
         android:layout_width="match_parent"
         android:layout_height="fill_parent"
         android:layout_below="@id/tab_layout"/>
-    <!--<FrameLayout-->
-    <!--android:id="@+id/fragmentContainer"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="match_parent">-->
-    <!--</FrameLayout>-->
-
-
-
-    <!--xmlns:app="http://schemas.android.com/apk/res-auto"-->
-    <!--xmlns:tools="http://schemas.android.com/tools"-->
-    <!--android:id="@+id/main_layout"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="match_parent"-->
-    <!--tools:context="com.example.lmont.iceicebb.GameActivity">-->
-
-
-    <!--<android.support.design.widget.TabLayout-->
-    <!--android:id="@+id/tab_layout"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="wrap_content"-->
-    <!--android:background="?attr/colorPrimary"-->
-    <!--android:elevation="8dp"-->
-    <!--android:minHeight="?attr/actionBarSize"-->
-    <!--/>-->
-
-    <!--<android.support.v4.view.ViewPager-->
-    <!--android:id="@+id/pager"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="fill_parent"-->
-    <!--android:layout_below="@id/tab_layout"/>-->
-
-    <!--<FrameLayout-->
-    <!--android:id="@+id/fragmentContainer"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="match_parent">-->
-    <!--</FrameLayout>-->
 
 
 </RelativeLayout>
