+++ IceIceBB/app/src/main/res/layout/fragment_feeling_lucky.xml
@@ -1,15 +1,24 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-             xmlns:tools="http://schemas.android.com/tools"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent"
-             tools:context="com.example.lmont.iceicebb.Fragments.FeelingLuckyFragment">
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                                             xmlns:app="http://schemas.android.com/apk/res-auto"
+                                             xmlns:tools="http://schemas.android.com/tools"
+                                             android:layout_width="match_parent"
+                                             android:layout_height="match_parent"
+                                             tools:context="com.example.lmont.iceicebb.Fragments.FeelingLuckyFragment">
 
 
-    <TextView
-        android:id="@+id/feeling_lucky"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-      android:textSize="40sp"/>
 
-</FrameLayout>
+
+    <Button
+        android:text="Feeling Lucky ?"
+        android:id="@+id/feelingLucky"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        android:layout_width="250dp"
+        android:layout_height="75dp"
+        android:layout_marginTop="16dp"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginBottom="16dp"
+        app:layout_constraintVertical_bias="0.96"/>
+
+</android.support.constraint.ConstraintLayout>
