+++ IceIceBB/app/src/main/res/layout/fragment_games.xml
@@ -4,6 +4,11 @@
              android:layout_height="match_parent"
              tools:context="com.example.lmont.iceicebb.Fragments.GamesFragment">
 
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/gameList"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_centerHorizontal="true"/>
 
     <TextView
         android:id="@+id/games"
