+++ IceIceBB/app/src/main/res/values/dimens.xml
@@ -6,5 +6,7 @@
     <dimen name="cardBoarderTopBottom">6dp</dimen>
     <dimen name="small_spacing">5dp</dimen>
     <dimen name="large_spacing">10dp</dimen>
+    <dimen name="fab_margin">16dp</dimen>
+    <dimen name="appbar_padding_top">8dp</dimen>
 
 </resources>
